dom/tv/test/mochitest/test_tv_scan_channels_completed.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test channel scanning complete for TV API</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript" src="./head.js"></script>
<script type="application/javascript">

function runTest() {
  ok('tv' in navigator, "navigator.tv should exist.");

  var isScannedEventFired = false;
  
  navigator.tv.getTuners().then(
    function(aTuners) {
      ok(aTuners.length > 0, "Got at least 1 tuner.");

      aTuners[0].getSources().then(
        function(aSources) {
          ok(aSources.length > 0, "Got at least 1 source.");
          var source = aSources[0];

          source.oneitbroadcasted = function(aEvent) {
            info("Received EIT broadcasted event.");

            var programs = aEvent.programs;
            for (var i = 0; i < programs.length; i++) {
              ok(programs[i], "Program " + i + " should be set.")
            }
          };

          source.onscanningstatechanged = function(aEvent) {
            if (aEvent.state === 'scanned') {
              isScannedEventFired = true;
              info("Received channel scanned event.");
              ok(aEvent.channel, "Scanned channel should be set.");
            } else if (aEvent.state === 'completed') {
              ok(isScannedEventFired, "Received channel scanning completed event after channel scanned event.");
              SimpleTest.finish();
            }
          };

          // TODO Bug 1088818 - Modify the behavior of channel scanning.
          source.startScanning({}).then(
            function() {},
            function(aError) {
              ok(false, "Error occurred when starting scanning: " + aError);
              SimpleTest.finish();
            }
          );
        },
        function(aError) {
          ok(false, "Error occurred when getting sources: " + aError);
          SimpleTest.finish();
        }
      );
    },
    function(aError) {
      ok(false, "Error occurred when getting tuners: " + aError);
      SimpleTest.finish();
    }
  );
}

SimpleTest.expectAssertions(0, 2);
SimpleTest.waitForExplicitFinish();
prepareTest(runTest);

</script>
</pre>
</body>
</html>
